Move `includeShebang` to `postprocess`
