+++ src/language-js/parser-flow.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const createError = require("../common/parser-create-error");
-const includeShebang = require("../common/parser-include-shebang");
 const { hasPragma } = require("./pragma");
 const locFns = require("./loc");
 const postprocess = require("./postprocess");
@@ -28,7 +27,6 @@ function parse(text, parsers, opts) {
     });
   }
 
-  includeShebang(text, ast);
   return postprocess(ast, { ...opts, originalText: text });
 }
