+++ src/language-js/parser-typescript.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const createError = require("../common/parser-create-error");
-const includeShebang = require("../common/parser-include-shebang");
 const { hasPragma } = require("./pragma");
 const locFns = require("./loc");
 const postprocess = require("./postprocess");
@@ -30,7 +29,6 @@ function parse(text, parsers, opts) {
     }
   }
 
-  includeShebang(text, ast);
   return postprocess(ast, { ...opts, originalText: text });
 }
