+++ src/language-js/postprocess.js
@@ -3,11 +3,16 @@
 const {
   getLast,
   getNextNonSpaceNonCommentCharacter,
+  getShebang
 } = require("../common/util");
 const { composeLoc, locEnd } = require("./loc");
 const { isTypeCastComment } = require("./comments");
 
 function postprocess(ast, options) {
+  if (options.parser === "typescript" || options.parser === "flow") {
+    includeShebang(ast, options);
+  }
+
   // Keep Babel's non-standard ParenthesizedExpression nodes only if they have Closure-style type cast comments.
   if (options.parser !== "typescript" && options.parser !== "flow") {
     const startOffsetsOfTypeCastedNodes = new Set();
@@ -180,6 +185,34 @@ function rebalanceLogicalTree(node) {
     right: node.right.right,
     ...composeLoc(node),
   });
+}
+
+
+function includeShebang(ast, options) {
+  const shebang = getShebang(options.originalText);
+
+  if (!shebang) {
+    return;
+  }
+  const index = shebang.length;
+
+  const comment = {
+    type: "Line",
+    value: shebang.slice(2),
+    range: [0, index],
+    loc: {
+      start: {
+        line: 1,
+        column: 0,
+      },
+      end: {
+        line: 1,
+        column: index,
+      },
+    },
+  };
+
+  ast.comments = [comment, ...ast.comments];
 }
 
 module.exports = postprocess;
